Revert "index was renamed"
