+++ pages/Employer/CompanyProfile/index.js
@@ -71,7 +71,7 @@ import { ERROR, SECTORS, SUCCESS } from "@/utils/constants";
 import { openAlert } from "@/redux/slices/alert";
 import { useRouter } from "next/navigation";
 import Location from "@/components/Location";
-import Employer from "../layout1";
+import Employer from "..";
 import { useTheme } from "@mui/material/styles";
 import { isEmpty } from "lodash";
 import companyservice from "@/redux/services/company.service";
@@ -236,11 +236,13 @@ const CompanyProfile = () => {
     if (e.target.value === "") {
       return;
     }
-    dispatch(updateFinaldetails(final)).then((res) => {
-      console.log(res, "res");
-      if (res?.meta?.requestStatus === "fulfilled") {
-        dispatch(openAlert({ type: SUCCESS, message: msg }));
-      }
+    dispatch(updateFinaldetails(final)).then((res)=>{
+console.log(res,'res')
+if (res?.meta?.requestStatus === "fulfilled") {
+  dispatch(
+    openAlert({ type: SUCCESS, message:msg })
+  )
+}
     });
   };
   const handleBlurDes = (value) => {
