+++ pages/Employer/Dashboard/index.js
@@ -51,13 +51,7 @@ import {
   getCompanyDetails,
   getapplCount,
 } from "@/redux/slices/companyslice";
-import {
-  companyJobs,
-  getfreeCount,
-  jobId,
-  setEditJob,
-  upgradejob,
-} from "@/redux/slices/job";
+import { companyJobs, getfreeCount, jobId, setEditJob, upgradejob } from "@/redux/slices/job";
 import { getSchedules } from "@/redux/slices/interviewslice";
 import moment from "moment";
 import { capitalizeFirstLetter } from "@/utils/HelperFunctions";
@@ -65,7 +59,7 @@ import axios from "axios";
 import { openAlert } from "@/redux/slices/alert";
 import { ERROR, SUCCESS } from "@/utils/constants";
 import dynamic from "next/dynamic";
-import Employer from "../layout1";
+import Employer from "..";
 
 const DatagridClient = dynamic(
   () => import("../../../components/Employers/DatagridClient"),
@@ -88,6 +82,7 @@ const BasicButton = styled(Button)({
   // width: "50%",
   marginBottom: "10px",
   marginTop: "15px",
+  
 });
 const BasicButtonUp = styled(Button)({
   color: "#ffff",
@@ -99,7 +94,7 @@ const BasicButtonUp = styled(Button)({
   marginBottom: "10px",
   marginTop: "15px",
   boxShadow:
-    "rgba(0, 0, 0, 0.25) 0px 14px 28px, rgba(0, 0, 0, 0.22) 0px 10px 10px",
+    "rgba(0, 0, 0, 0.25) 0px 14px 28px, rgba(0, 0, 0, 0.22) 0px 10px 10px"
 });
 
 const useStyles = makeStyles((theme) => ({
@@ -114,11 +109,9 @@ const useStyles = makeStyles((theme) => ({
     minWidth: 190,
   },
 }));
-
 const Transition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="left" ref={ref} {...props} />;
 });
-
 const TabPanel = (props) => {
   const { children, value, index, ...other } = props;
 
@@ -144,14 +137,12 @@ TabPanel.propTypes = {
   index: PropTypes.number.isRequired,
   value: PropTypes.number.isRequired,
 };
-
 function a11yProps(index) {
   return {
     id: `simple-tab-${index}`,
     "aria-controls": `simple-tabpanel-${index}`,
   };
 }
-
 const EmpoyerDashboard = () => {
   let dispatch = useDispatch();
 
@@ -527,7 +518,7 @@ const EmpoyerDashboard = () => {
     dispatch(upgradejob([parms.id]));
     push("/Employer/Jobpayment");
   };
-
+  
   const columns = [
     { field: "id", headerName: "Id", width: 100, hide: true },
     { field: "_id", headerName: "Job", width: 120 },
@@ -779,22 +770,16 @@ const EmpoyerDashboard = () => {
       width: 200,
       renderCell: (parms) => (
         <>
-          <Button
-            sx={{ textTransform: "capitalize", color: "green" }}
-            onClick={() => {
-              handleClose2();
-              handleUpgrade(parms?.row);
-            }}
-          >
-            Upgrade To Premium
-          </Button>
+          <Button sx={{textTransform:'capitalize',color:'green'}}  onClick={() => {
+                    handleClose2();
+                    handleUpgrade(parms?.row);
+                  }}>
+                    Upgrade To Premium</Button>
         </>
       ),
     },
   ];
-  const updatedColumns = columns.filter(
-    (column) => column.headerName !== "Upgrade"
-  );
+  const updatedColumns = columns.filter((column) => column.headerName !== 'Upgrade')
 
   const handleGetRowId = (e) => {
     return e.id;
@@ -806,10 +791,10 @@ const EmpoyerDashboard = () => {
     // eslint-disable-next-line no-mixed-operators
     company?.package?.subscription_package === "SuperEmployer" ||
     company?.jobCounts?.premiumPayment === "Completed" ||
-    company?.package?.subscription_package === "Gold" ||
+    company?.package?.subscription_package === "Gold" || 
     company?.jobSlotGold === true ||
-    user?.country === "IN";
-
+    user?.country === 'IN'
+    
   const enableFeaturedJobs2 =
     // eslint-disable-next-line no-mixed-operators
     company?.package?.subscription_package === "SuperEmployer" ||
@@ -843,7 +828,6 @@ const EmpoyerDashboard = () => {
     }
     setOpen1(false);
   };
-
   const [freejobs, setfreejobs] = useState([]);
   const selectJobs = (e) => {
     const { value, checked } = e.target;
@@ -1418,23 +1402,15 @@ const EmpoyerDashboard = () => {
             </Card>
           </Grid>
         </Grid>
-        {user.country !== "LK" ? (
-          sectors?.length > 0 ? (
-            <Box sx={{ display: "flex", justifyContent: "flex-end" }}>
-              <BasicButtonUp
-                onClick={() => {
-                  handleClickOpen();
-                }}
-              >
-                Upgrade To Premium
-              </BasicButtonUp>
-            </Box>
-          ) : (
-            ""
-          )
-        ) : (
-          ""
-        )}
+   {user.country !== 'LK' ?  sectors?.length > 0 ?<Box sx={{ display: "flex", justifyContent: "flex-end", }}>
+          <BasicButtonUp
+            onClick={() => {
+              handleClickOpen();
+            }}
+          >
+            Upgrade To Premium
+          </BasicButtonUp>
+        </Box> : ""  : ''}
         {user?.country === "LK" ? (
           <Stack
             direction={{ xs: "column", sm: "row" }}
@@ -1548,7 +1524,7 @@ const EmpoyerDashboard = () => {
                 sx={{ display: "flex", justifyContent: "center" }}
                 getRowId={handleGetRowId}
                 rows={rows}
-                columns={user?.country === "LK" ? updatedColumns : columns}
+                columns={user?.country === 'LK' ? updatedColumns : columns}
               />
             </Box>
           </TabPanel>
@@ -1603,7 +1579,7 @@ const EmpoyerDashboard = () => {
                   {sectors?.map((sec, index) => (
                     <FormControlLabel
                       key={index}
-                      sx={{ width: "fit-content" }}
+                      sx={{width:'fit-content'}}
                       control={
                         <Checkbox
                           checked={freejobs.includes(sec?._id)}
@@ -1658,3 +1634,4 @@ const EmpoyerDashboard = () => {
   );
 };
 export default dynamic(() => Promise.resolve(EmpoyerDashboard), { ssr: false });
+
