+++ pages/Employer/MyAccount/index.js
@@ -25,7 +25,7 @@ import {
   IconButton,
 } from "@mui/material";
 import { Visibility, VisibilityOff } from "@mui/icons-material";
-import Employer from "../layout1";
+import Employer from "..";
 import { useDispatch } from "react-redux";
 import http from "@/redux/http-common";
 import { openAlert } from "@/redux/slices/alert";
